+++ packages/navigation/app-navigation/src/common/styles.ts
@@ -1,0 +1,14 @@
+import { css } from '@emotion/core';
+import { CREATE_BREAKPOINT } from './constants';
+
+export const actionSectionDesktopStyles = css`
+  @media (max-width: ${CREATE_BREAKPOINT - 1}px) {
+    display: none;
+  }
+`;
+
+export const actionSectionMobileStyles = css`
+  @media (min-width: ${CREATE_BREAKPOINT}px) {
+    display: none;
+  }
+`;