+++ packages/navigation/app-navigation/src/components/Settings/types.ts
@@ -1,0 +1,5 @@
+import { TriggerManagerProps } from '../TriggerManager/types';
+
+export type SettingsProps = Omit<TriggerManagerProps, 'children'> & {
+  tooltip: string;
+};