+++ packages/navigation/atlassian-switcher/examples/01-theme-builder.tsx
@@ -1,0 +1,94 @@
+import * as React from 'react';
+import { mockEndpoints, REQUEST_FAST } from './helpers/mock-endpoints';
+import { withAnalyticsLogger, withIntlProvider } from './helpers';
+import AtlassianSwitcher from '../src';
+import ThemeBuilder from './helpers/ThemeBuilder';
+import styled from 'styled-components';
+
+const Container = styled.div`
+  width: 300px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  padding: 0 8px;
+  display: inline-block;
+  margin: 5px;
+  vertical-align: top;
+`;
+class InlineDialogSwitcherExample extends React.Component {
+  state = {
+    isLoaded: false,
+  };
+
+  componentDidMount() {
+    this.loadData();
+  }
+  loadData = () => {
+    mockEndpoints(
+      'jira',
+      originalMockData => {
+        return {
+          ...originalMockData,
+          RECENT_CONTAINERS_DATA: {
+            data: [],
+          },
+          CUSTOM_LINKS_DATA: {
+            data: [],
+          },
+          XFLOW_SETTINGS: {},
+
+          LICENSE_INFORMATION_DATA: {
+            hostname: 'https://some-random-instance.atlassian.net',
+            firstActivationDate: 1492488658539,
+            maintenanceEndDate: '2017-04-24',
+            maintenanceStartDate: '2017-04-17',
+            products: {
+              'jira-software.ondemand': {
+                billingPeriod: 'ANNUAL',
+                state: 'ACTIVE',
+              },
+            },
+          },
+        };
+      },
+      REQUEST_FAST,
+    );
+    this.setState({
+      isLoaded: true,
+    });
+  };
+
+  onTriggerXFlow = (productKey: string, sourceComponent: string) => {
+    console.log(
+      `Triggering xflow for => ${productKey} from ${sourceComponent}`,
+    );
+  };
+
+  render() {
+    return (
+      this.state.isLoaded && (
+        <ThemeBuilder>
+          {theme => (
+            <Container>
+              <AtlassianSwitcher
+                product="trello"
+                disableCustomLinks
+                disableRecentContainers
+                disableHeadings
+                isDiscoverMoreForEveryoneEnabled
+                enableUserCentricProducts
+                cloudId="some-cloud-id"
+                triggerXFlow={this.onTriggerXFlow}
+                appearance="standalone"
+                theme={theme}
+              />
+            </Container>
+          )}
+        </ThemeBuilder>
+      )
+    );
+  }
+}
+
+export default withIntlProvider(
+  withAnalyticsLogger(InlineDialogSwitcherExample),
+);